@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   '*.{js,jsx,ts,tsx}': ['eslint --fix', 'eslint'],
   '**/*.ts?(x)': () => 'npm run check-types',
   '*.json': ['prettier --write'],
